Converted amount field of saved wealths to double

+++ wealth_calculator/lib/feature/inventory/model/wealths_model.dart
@@ -2,7 +2,7 @@
 class SavedWealths {
   final int id;
   final String type;
-  late final int amount;
+  late final double amount;
 
   SavedWealths({required this.id, required this.type, required this.amount});
